+++ Automation/ScriptingEngine/ScriptableComponents/Extensions/Configurator.cs
@@ -1,0 +1,15 @@
+﻿// Timberborn Mod: Automation
+// Author: igor.zavoychinskiy@gmail.com
+// License: Public Domain
+
+using Bindito.Core;
+
+namespace IgorZ.Automation.ScriptingEngine.ScriptableComponents.Extensions;
+
+// ReSharper disable once UnusedType.Global
+[Context("Game")]
+sealed class Configurator : IConfigurator {
+  public void Configure(IContainerDefinition containerDefinition) {
+    containerDefinition.Bind<AutomationExtensionsRegistry>().AsSingleton();
+  }
+}