+++ Automation/ScriptingEngine/ScriptableComponents/Extensions/IAutomationExtension.cs
@@ -1,0 +1,10 @@
+﻿// Timberborn Mod: Automation
+// Author: igor.zavoychinskiy@gmail.com
+// License: Public Domain
+
+namespace IgorZ.Automation.ScriptingEngine.ScriptableComponents.Extensions;
+
+/// <summary>Marker interface for automation extensions.</summary>
+/// <seealso cref="AutomationExtensionsRegistry"/>
+public interface IAutomationExtension {
+}