+++ Automation/CHANGELOG.md
@@ -4,6 +4,8 @@
 * [Feature] Add `Plantable.Ready` signal that tells how many plantable resources can be planted in the range.
 * [Feature #82] Add templates to pause buildings if they can't gather or cut anything.
 * [Feature] Add a template to pause buildings if they can't plant trees or crops.
+* [Feature] Implemented extensions support to allow other mods to modify rules options.
+* [Change] `ModdedWeather` mod support moved into a separate extension mod: `Automation+ModdedWeather`.
 * [Fix] In "describe mode," show action argument values as real numbers instead of script fixed point values.
 
 # v2.6.0 (June 22nd, 2025):